[UIMA-5922] remove unnecessary boxing / unboxing
git-svn-id: https://svn.apache.org/repos/asf/uima/uv3/uimaj-v3/trunk@1848164 13f79535-47bb-0310-9956-ffa450edef68

+++ uimaj-adapter-vinci/src/main/java/org/apache/uima/adapter/vinci/VinciAnalysisEngineServiceStub.java
@@ -146,7 +146,7 @@ public class VinciAnalysisEngineServiceStub implements AnalysisEngineServiceStub
       
       //store timeout for use in later RPC calls
       if (timeout != null) {
-        mTimeout = timeout.intValue();
+        mTimeout = timeout;
       } else {
        mTimeout = mVinciClient.getSocketTimeout(); //default
       }
