+++ uimaj-core/src/main/java/org/apache/uima/cas/impl/FeatureValuePathImpl.java
@@ -269,7 +269,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 			switch (typeClass) {
 			case LowLevelCAS.TYPE_CLASS_FLOATARRAY:
 				int position = getArrayIndex(arraySize);
-				return Float.valueOf(cas.ll_getFloatArrayValue(arrayFS, position, false));
+				return cas.ll_getFloatArrayValue(arrayFS, position, false);
 			case LowLevelCAS.TYPE_CLASS_FSARRAY:
 				int childFS = getFsAtIndex(arrayFS, cas, arraySize);
 				return this.childPath.evaluateAsFloat(childFS, cas);
@@ -309,7 +309,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 			int typeClass = cas.ll_getTypeClass(this.featureRangeType);
 			switch (typeClass) {
 			case LowLevelCAS.TYPE_CLASS_FLOAT:
-				return Float.valueOf(cas.ll_getFloatValue(currentFS, this.featureCode, CAS_TYPE_CHECKS));
+				return cas.ll_getFloatValue(currentFS, this.featureCode, CAS_TYPE_CHECKS);
 			default:
 				throw new IllegalStateException("feature path does not denote a float");
 			}
@@ -356,7 +356,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 				} else {
 					// this snippet denotes a float array, just collect it
 					for (int i = 0; i < arraySize; i++) {
-						result[i] = Float.valueOf(cas.ll_getFloatArrayValue(arrayFS, i));
+						result[i] = cas.ll_getFloatArrayValue(arrayFS, i);
 					}
 				}
 				return result;
@@ -386,7 +386,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 					// FSList
 					// iterate through the results , which will return floats
 					for (int i = 0; i < resultList.size(); i++) {
-						int childFS = ((Integer) resultList.get(i)).intValue();
+						int childFS = (Integer) resultList.get(i);
 						result[i] = this.childPath.evaluateAsFloat(childFS, cas);
 					}
 				} else {
@@ -428,7 +428,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 			switch (typeClass) {
 			case LowLevelCAS.TYPE_CLASS_INTARRAY:
 				int position = getArrayIndex(arraySize);
-				return Integer.valueOf(cas.ll_getIntArrayValue(arrayFS, position, false));
+				return cas.ll_getIntArrayValue(arrayFS, position, false);
 			case LowLevelCAS.TYPE_CLASS_FSARRAY:
 				int childFS = getFsAtIndex(arrayFS, cas, arraySize);
 				return this.childPath.evaluateAsInt(childFS, cas);
@@ -464,14 +464,14 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 		} else if (this.isCoveredTextFeature || this.isTypeNameFeature) {
 			throw new IllegalStateException("feature path does not denote an int");
 		} else if (this.isFsIdFeature) {
-			return Integer.valueOf(currentFS);
+			return currentFS;
 		} else if (this.isUniqueIdFeature) {
-			return Integer.valueOf(currentFS); // TODO: return currentFs + chunkId
+			return currentFS; // TODO: return currentFs + chunkId
 		} else {
 			int typeClass = cas.ll_getTypeClass(this.featureRangeType);
 			switch (typeClass) {
 			case LowLevelCAS.TYPE_CLASS_INT:
-				return Integer.valueOf(cas.ll_getIntValue(currentFS, this.featureCode, CAS_TYPE_CHECKS));
+				return cas.ll_getIntValue(currentFS, this.featureCode, CAS_TYPE_CHECKS);
 			default:
 				throw new IllegalStateException("feature path does not denote an int");
 			}
@@ -518,7 +518,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 				} else {
 					// this snippet denotes an int array, just collect it
 					for (int i = 0; i < arraySize; i++) {
-						result[i] = Integer.valueOf(cas.ll_getIntArrayValue(arrayFS, i));
+						result[i] = cas.ll_getIntArrayValue(arrayFS, i);
 					}
 				}
 				return result;
@@ -548,7 +548,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 					// FSList
 					// iterate through the results , which will return floats
 					for (int i = 0; i < resultList.size(); i++) {
-						int childFS = ((Integer) resultList.get(i)).intValue();
+						int childFS = (Integer) resultList.get(i);
 						result[i] = this.childPath.evaluateAsInt(childFS, cas);
 					}
 				} else {
@@ -731,7 +731,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 					// FSList
 					// iterate through the results , which will return floats
 					for (int i = 0; i < resultList.size(); i++) {
-						int childFS = ((Integer) resultList.get(i)).intValue();
+						int childFS = (Integer) resultList.get(i);
 						result[i] = this.childPath.evaluateAsString(childFS, cas);
 					}
 				} else {
@@ -1077,7 +1077,7 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 		}
 		Object valueAtListIndex = getValueAtListIndex(cas, listFS);
 		if (valueAtListIndex != null) {
-			return ((Integer) valueAtListIndex).intValue();
+			return (Integer) valueAtListIndex;
 		}
 		return 0;
 	}
@@ -1090,11 +1090,11 @@ public class FeatureValuePathImpl implements FeatureValuePath {
 		case TYPE_CLASS_STRINGLIST:
 			return cas.ll_getStringValue(listFS, this.headFeature);
 		case TYPE_CLASS_INTEGERLIST:
-			return Integer.valueOf(cas.ll_getIntValue(listFS, this.headFeature));
+			return cas.ll_getIntValue(listFS, this.headFeature);
 		case TYPE_CLASS_FLOATLIST:
-			return Float.valueOf(cas.ll_getFloatValue(listFS, this.headFeature));
+			return cas.ll_getFloatValue(listFS, this.headFeature);
 		case TYPE_CLASS_FSLIST:
-			return Integer.valueOf(cas.ll_getRefValue(listFS, this.headFeature));
+			return cas.ll_getRefValue(listFS, this.headFeature);
 		default:
 			return null;
 		}
