+++ uimaj-core/src/test/java/org/apache/uima/cas/impl/FeaturePathTest.java
@@ -107,7 +107,7 @@ public class FeaturePathTest extends TestCase {
       featurePath.initialize(path);
       assertEquals(path, featurePath.getFeaturePath());
       featurePath.typeInit(cas.getDocumentAnnotation().getType());
-      assertEquals(Float.valueOf(1.12f), featurePath.getFloatValue(cas
+      assertEquals(1.12f, featurePath.getFloatValue(cas
             .getDocumentAnnotation()));
       assertEquals("1.12", featurePath.getValueAsString(cas
             .getDocumentAnnotation()));
@@ -127,7 +127,7 @@ public class FeaturePathTest extends TestCase {
       featurePath.initialize(path);
       assertEquals(path, featurePath.getFeaturePath());
       featurePath.typeInit(cas.getDocumentAnnotation().getType());
-      assertEquals(Double.valueOf(100.5), featurePath.getDoubleValue(cas
+      assertEquals(100.5, featurePath.getDoubleValue(cas
             .getDocumentAnnotation()));
       assertEquals("100.5", featurePath.getValueAsString(cas
             .getDocumentAnnotation()));
